Added tableExists function

+++ dostuff.py
@@ -1,5 +1,9 @@
 import sqlite3
 
+# Checks if Table already Exists in DB
+def tableExists(tableName):
+    return len(cursor.execute(f"SELECT name FROM sqlite_master WHERE type='table' AND name='{tableName}'").fetchall()) != 0
+
 connection = sqlite3.connect("practice.db")
 cursor = connection.cursor()
 while True:
@@ -10,14 +14,14 @@ while True:
         case("CREATE TABLE"):
             try:
                 tableName = input("Input table name: ")
-                if(len(cursor.execute(f"SELECT name FROM sqlite_master WHERE type='table' AND name='{tableName}'").fetchall()) != 0):
+                if(tableExists(tableName)):
                     print(f"{tableName} Already Exists")
                     break
-                # Get User input for amount of columns wanted in table
+                # Get User Input for amount of columns wanted in table
                 while True:
                     try:
                         columnAmount = int(input("Input total number of columns: "))
-                        if columnAmount < 0:
+                        if columnAmount <= 0:
                             raise ValueError
                         break
                     except ValueError:
@@ -38,6 +42,10 @@ while True:
         case("DELETE TABLE"):
             try:
                 tableName = input("Input Table Name: ")
+                # Checks if Table Exists
+                if(not tableExists(tableName)):
+                    print(f"{tableName} Doesn't Exists")
+                    break
                 cursor.execute(f"DROP TABLE {tableName}")
                 connection.commit()
                 print("Succesfully Deleted Table")
@@ -47,32 +55,41 @@ while True:
         case("INSERT"):
             try:
                 tableName = input("Input Table Name: ")
+                if(not tableExists(tableName)):
+                    print(f"{tableName} Doesn't Exists")
+                    break
                 pragma = cursor.execute(f"PRAGMA table_info({tableName})").fetchall()
                 columnNum = 0
+                #Prints out Column'ss Name and Type
                 for column in pragma:
                     columnNum += 1
                     print(f"Column {columnNum}: {list(column)[1]} {list(column)[2]}")
+
                 columnValues = []
                 sanitizerList = []
                 for x in range(columnNum):
                     currentColumnNumber = x+1
                     columnValues.append(input(f"Input Column {currentColumnNumber} Value: "))
                     sanitizerList.append("?")
-                sanitizer = ",".join(sanitizerList)
-                sanitizerTuple = sanitizer
-                cursor.execute(f"INSERT INTO {tableName} VALUES({sanitizerTuple})",(tuple(columnValues)))
+                sanitizer = ",".join(sanitizerList) # Contains "?" of Column Total for sanitization
+                cursor.execute(f"INSERT INTO {tableName} VALUES({sanitizer})",tuple(columnValues))
                 connection.commit()
+                print(f"Successfully Inserted Element into {tableName}")
             except Exception as e:
                 print(e)
                 print("Failed to Insert Element")
         case("DELETE"):
             try:
                 tableName = input("Input Table Name: ")
+                if(not tableExists(tableName)):
+                    print(f"{tableName} Doesn't Exists")
+                    break
             except Exception as e:
                 print(e)
-                print("Failed to Delete Element")
+                print("Failed to Delete Table")
         case _:
             print("Invalid Command")
 
-
 connection.close()
+
+
